% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_thresh_wavelet_coeff.R
\name{apply_thresh_wavelet_coeff}
\alias{apply_thresh_wavelet_coeff}
\title{Function to apply thresholding on warped wavelet coefficients.
The thresholding steps are based on the estimator in Eq. (16) of
Kerkyacharian and Picard (2004, Bernoulli).}
\usage{
apply_thresh_wavelet_coeff(
  x = NULL,
  y = NULL,
  wm_filter_num = NULL,
  wm_family = NULL,
  threshold_type = NULL,
  thresh_par = NULL,
  n_levels = NULL,
  boundary_handling = "periodic"
)
}
\arguments{
\item{x}{a vector with observations of the covariate/feature.}

\item{y}{a vector with observations of the response variable.}

\item{wm_filter_num}{the filter.number variable in the wavethresh::wd
function. The higher it is, the smoother the fitted functions are.}

\item{wm_family}{the wavelet family used in the wavethresh::wd function.}

\item{thresh_par}{tuning parameter used in the threshold of wavelet
coefficients. The threshold value is thresh_par * sqrt(log(n)/n).}

\item{n_levels}{number of levels used in the warped wavelet estimator.}

\item{boundary_handling}{boundary handling used in wavethresh 'bc' argument.}
}
\value{
thresholded wavelet coefficients
}
\description{
Function to apply thresholding on warped wavelet coefficients.
The thresholding steps are based on the estimator in Eq. (16) of
Kerkyacharian and Picard (2004, Bernoulli).
}
