% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_warped_wavelet_coefficients.R
\name{compute_warped_wavelet_coefficients}
\alias{compute_warped_wavelet_coefficients}
\title{Computes warped wavelet coefficients and then apply thresholding
on them. The computed estimator is presented in Eq. (16) of
Kerkyacharian and Picard (2004, Bernoulli).}
\usage{
compute_warped_wavelet_coefficients(
  x = NULL,
  y = NULL,
  wm_filter_num = 1,
  wm_family = "DaubExPhase",
  thresh_type = "hard",
  thresh_par = NULL,
  n_levels = NULL,
  boundary_handling = "periodic"
)
}
\arguments{
\item{x}{a vector with observations of the covariate/feature.}

\item{y}{a vector with observations of the response variable.}

\item{wm_filter_num}{the filter.number variable in the wavethresh::wd
function. The higher it is, the smoother the fitted functions are.}

\item{wm_family}{the wavelet family used in the wavethresh::wd function.}

\item{thresh_par}{tuning parameter used in the threshold of wavelet
coefficients. The threshold value is thresh_par * sqrt(log(n)/n).}

\item{n_levels}{number of levels used in the warped wavelet estimator.}

\item{boundary_handling}{boundary handling used in wavethresh 'bc' argument.}
}
\value{
A list with three elemets:
wavelet_coeff_thr: thresholded wavelet coefficients;
n_first_extend_values: number of observations appended at the beginning of
the vector if it does not have a length that is a power of two;
n_last_extend_values: number of observations appended at the end of
the vector if it does not have a length that is a power of two.
}
\description{
Computes warped wavelet coefficients and then apply thresholding
on them. The computed estimator is presented in Eq. (16) of
Kerkyacharian and Picard (2004, Bernoulli).
}
