% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extend_vector_length_topower2.R
\name{extend_vector_length_topower2}
\alias{extend_vector_length_topower2}
\title{Function that extends a vector to make it have a length that is a power of
two. The extension is done via a symmetric repetition of values in the
border. This function is similar to "wextend('1D','sym',x,len)" in Matlab.}
\usage{
extend_vector_length_topower2(y = null)
}
\arguments{
\item{y}{vector to be extended.}
}
\value{
A list with three elemets:
y_extended: the extended vector;
n_first_extend_values: number of repeated values in the beginning of y_extended;
n_last_extend_values: number of repeated values in the end of y_extended.
}
\description{
Function that extends a vector to make it have a length that is a power of
two. The extension is done via a symmetric repetition of values in the
border. This function is similar to "wextend('1D','sym',x,len)" in Matlab.
}
\examples{
y = c(1:9)
extend_vector_length_topower2(y)
y = c(1:7)
extend_vector_length_topower2(y)

}
