% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wavelet_screening.R
\name{wavelet_screening}
\alias{wavelet_screening}
\title{Feature screening via warped wavelets.}
\usage{
wavelet_screening(
  x,
  y,
  wm_filter_num = 1,
  wm_family = "DaubExPhase",
  thresh_type = "hard",
  thresh_par = 1,
  n_levels = 1,
  boundary_handling = "periodic"
)
}
\arguments{
\item{x}{nxp matrix of covariates.}

\item{y}{n-dimensional vector with response variable.}

\item{wm_filter_num}{the filter.number variable in the wavethresh::wd
function. The higher it is, the smoother the fitted functions are.}

\item{wm_family}{the wavelet family used in the wavethresh::wd function.}

\item{thresh_par}{tuning parameter used in the threshold of wavelet
coefficients. The threshold value is thresh_par * sqrt(log(n)/n).}

\item{n_levels}{number of levels used in the warped wavelet estimator.}

\item{boundary_handling}{boundary handling used in wavethresh 'bc' argument.}
}
\value{
A list with two elemets:
rank: importance of the variable (important variables are expected to
have the lowest indices);
measurement: the screening utility used to compute the rank.
}
\description{
Feature screening via warped wavelets.
}
\examples{
sample_size = 100
number_parameters = 500
matrix_features = matrix(data = runif(sample_size * number_parameters),
                         nrow=sample_size,
                         ncol=number_parameters)
response_variable = sin(2*pi*matrix_features[,1]) +
                      3*(matrix_features[,2]^2) + 0.5*rnorm(sample_size)
wws <- wavelet_screening(matrix_features, response_variable)
wws$rank[1:20]

}
